Move web_poet.testing content from __init__.py.
